make test nv16 specific
